Inline serverMode as true in WebSocketFrameParser
The core module's WebSocketFrameParser always runs in server mode
(both QwpWebSocketUpgradeProcessor and WebSocketConnectionContext
called setServerMode(true)). Remove the flag and inline the
server-mode behavior: always require masked frames from clients.

Remove setServerMode() and the client-mode code path. Update all
tests to use masked frames via createMaskedFrame(). Remove the
two client-specific tests (testRejectMaskedServerFrame,
testParseMinimalFrame) that now live in the client module.

Fix createMaskedFrame() 64-bit length encoding bug (int shift
wrapping). Fix WebSocketFrameWriterTest round-trip tests to use
masked frames.

+++ core/src/test/java/io/questdb/test/cutlass/http/websocket/AbstractWebSocketTest.java
@@ -75,7 +75,7 @@ public abstract class AbstractWebSocketTest extends AbstractTest {
         } else {
             frame[offset++] = (byte) (0x80 | 127);
             for (int i = 7; i >= 0; i--) {
-                frame[offset++] = (byte) ((payloadLen >> (i * 8)) & 0xFF);
+                frame[offset++] = (byte) (((long) payloadLen >> (i * 8)) & 0xFF);
             }
         }
